Add handling of the realm metadata
This change introduces a mechanism of passing an optional realm
metadata to a realm. The realm metadata block provides an authority
and identity information about the realm. This information can be
used to derive sealing keys that are immune to updates of the realm.

Merged with:
Change the order of arguments for the populate realm metadata smc

Signed-off-by: Piotr Sawicki <p.sawicki2@samsung.com>